Add refresh button and view state animations
Now using `AndroidViewModel` to make Android resources available to the
view model and to simplify view model creation.

Fixed OkHttp logging to add `HttpLoggingInterceptor` as a network
interceptor. It was suspected that the logger was consuming our response
objects before we could get to it.

Abandoning use of [Result] in favor of custom `sealed class` based
states.

[Result]: https://github.com/kittinunf/Result